requirements.txt
requests==2.31.0
python-dotenv==1.0.0
azure-storage-file-datalake==12.14.0
azure-identity==1.15.0
pyspark==3.5.0
apache-airflow==2.8.0
kafka-python==2.0.2
pyodbc==5.0.1
pandas==2.1.4
great-expectations==0.18.0
pytest==7.4.0
